module update

+++ tflearn_regression_experiment.py
@@ -1,7 +1,5 @@
 # -*- encoding: utf-8 -*-
 from __future__ import print_function
-import pandas as pd
-import numpy as np
 import os
 import datetime
 from tflearn_regression import SimpleModel
@@ -18,8 +16,8 @@ def simulate(bd, ed):
     sm.train_model_tensorflow(X_train, Y_train, s_date)
     sm.save_scaler(s_date)
     sm.load_scaler(s_date)
-    test_bd = datetime.date(ed/10000, ed%10000/100, ed%100) - datetime.timedelta(days=40)
-    test_ed = datetime.date(ed/10000, ed%10000/100, ed%100) + datetime.timedelta(days=40)
+    test_bd = datetime.date(int(ed/10000), int(ed%10000/100), ed%100) - datetime.timedelta(days=40)
+    test_ed = datetime.date(int(ed/10000), int(ed%10000/100), ed%100) + datetime.timedelta(days=40)
     test_bd = int(test_bd.strftime("%Y%m%d"))
     test_ed = int(test_ed.strftime("%Y%m%d"))
     print("Evaluation on %d - %d" % (test_bd, test_ed))
@@ -34,13 +32,12 @@ def simulate(bd, ed):
 def simulate_all():
     base_month = 201501
     while base_month <= 201701:
-        end_date = datetime.date(base_month/100, base_month%100, 1)
-        begin_date = datetime.date(base_month/100-5, base_month%100, 1)
+        end_date = datetime.date(int(base_month/100), base_month%100, 1)
+        begin_date = datetime.date(int(base_month/100)-5, base_month%100, 1)
         print(begin_date)
         p = Process(target=simulate, args=(begin_date.strftime("%Y%m%d"), end_date.strftime("%Y%m%d")))
         p.start()
         p.join()
-        #simulate(begin_date.strftime("%Y%m%d"), end_date.strftime("%Y%m%d"))
         base_month += 1
         if base_month%100 > 12:
             base_month += 88
